Add support for support forums (#14)
Looks at forum categories with "Support" in the name and adds the functionality described in issue #14.
We now filter out posts more aggressively, which makes one test (which checks if we return deleted posts) falsely error because, funnily enough, we don't return a deleted post.
(For context, that test gets sent a string indicating that a post is deleted, and checks for that string; since we never return that post at all, it errors...)
So, we fix that test accordingly.